requirement.txt
altair==5.3.0
matplotlib==3.8.3
numpy==1.26.4
opencv-python==4.9.0.80
streamlit==1.35.0
pillow==10.2.0
